LPS-4093, changing workflow definition version from double to int
git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@38848 05bdf26c-840f-0410-9ced-eb539d925f36

+++ webs/edoras-web/docroot/WEB-INF/service/com/liferay/portal/workflow/edoras/model/WorkflowDefinitionModel.java
@@ -74,9 +74,9 @@ public interface WorkflowDefinitionModel extends BaseModel<WorkflowDefinition> {
 
 	public void setName(String name);
 
-	public double getVersion();
+	public int getVersion();
 
-	public void setVersion(double version);
+	public void setVersion(int version);
 
 	public String getDesignerVersion();
